Accepted UVA10474
This problem is tricky, specially if the concept of Binary search is not
clear. Though this problem is a best case of Binary search but BS wont work
for all the test cases.

So a simple look up for element in a sorted array or list worked fine.
Though it is not the best solution of this problem.

Scanner or BufferedReader both can be used.
Also Array and List both can be used. (Though I used List)

Signed-off-by: imrul001 <imrul001@gmail.com>